feat : bug수정 및 reload
