+++ jakkufront/src/pages/Topbar.jsx
@@ -4,11 +4,12 @@ import Topbar_top from "../components/topbar/Topbar_top";
 import Topbar_bottom from "../components/topbar/Topbar_bottom";
 import ModuleStyle from "../ModuleStyle.module.css";
 import api from "../axios";
-import { nicknameState } from "../atoms/atom";
+import {nicknameState, usermoneyState} from "../atoms/atom";
 import {useRecoilState} from "recoil";
 
 function Topbar() {
     const [nickname, setNickname] = useRecoilState(nicknameState);
+    const [remainMoney, setRemainMoney] = useRecoilState(usermoneyState);
     const fetchNickname = async () => {
         try {
             const response = await api.get("/topbar/username");
@@ -18,8 +19,19 @@ function Topbar() {
         }
     };
 
+    const remainingMoney = async () => {
+        try{
+            const response = await api.get(`/customers/basket/point`);
+            console.log("남은 돈 확인");
+            setRemainMoney(response.data.responseDto);
+        } catch (error) {
+            console.error("Error fetching remaining money:", error);
+        }
+    }
+
     useEffect(() => {
         fetchNickname();
+        remainingMoney();
     }, []);
 
     return(
